backend/requirements.txt
Django
uWSGI
psycopg2-binary
djangorestframework
djangorestframework-simplejwt
djangorestframework_jwt
django-rest-auth
django-rest-auth[with_social]
django-cors-headers